requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
flake8==7.3.0
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
iniconfig==2.1.0
Markdown==3.8.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
uritemplate==4.2.0
